requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
greenlet==3.2.4
idna==3.11
iniconfig==2.1.0
loguru==0.7.3
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
psycopg-binary==3.2.11
pygments==2.19.2
pytest==8.4.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlalchemy==2.0.44
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
win32-setctime==1.2.0
